+++ src/wrap/opaque.rs
@@ -14,12 +14,7 @@ use derive_more::{AsMut, AsRef, Deref, From};
 /// # use display_as_debug::wrap::Opaque;
 /// # use display_as_debug::types::OPAQUE;
 /// assert_eq!(format!("{:?}", OPAQUE), "..", "Debug format should be opaque");
-/// assert_eq!(format!("{}", OPAQUE), "..", "Display format should be opaque");
-///
-/// assert_eq!(format!("{:?}", Opaque("secret")), "..", "Debug format should be opaque");
-/// assert_eq!(format!("{}", Opaque("secret")), "..", "Display format should be opaque");
-/// ```
-#[derive(Copy, Clone, Deref, From, AsMut, AsRef, Default)]
+#[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Deref, From, AsMut, AsRef, Default)]
 pub struct Opaque<T = ()>(pub T);
 
 impl Opaque<()> {