+++ tests/types/type_name/mod.rs
@@ -2,6 +2,7 @@
 
 mod list;
 mod option;
+mod result;
 mod set;
 #[allow(clippy::module_inception, reason = "Matches implementation module name")]
 mod type_name;